+++ lib/model.g.dart
@@ -335,6 +335,7 @@ const _$RPConsentSectionTypeEnumMap = {
   RPConsentSectionType.Benefits: 'Benefits',
   RPConsentSectionType.DataHandling: 'DataHandling',
   RPConsentSectionType.Location: 'Location',
+  RPConsentSectionType.BackgroundSensing: 'BackgroundSensing',
   RPConsentSectionType.Health: 'Health',
   RPConsentSectionType.HealthDataCollection: 'HealthDataCollection',
   RPConsentSectionType.UserDataCollection: 'UserDataCollection',
@@ -351,6 +352,7 @@ const _$RPPermissionTypeEnumMap = {
   RPPermissionType.activityRecognition: 'activityRecognition',
   RPPermissionType.sensors: 'sensors',
   RPPermissionType.bluetooth: 'bluetooth',
+  RPPermissionType.ignoreBatteryOptimizations: 'ignoreBatteryOptimizations',
   RPPermissionType.health: 'health',
 };